Fix unsafe use of symbolic links in /tmp
If a script uses a temp file which is created in /tmp, then an
attacker can create symlink with the same name in this directory in
order to destroy or rewrite some system or user files.  Symlink attack
may also lead not only to the data desctruction but to denial of
service as well. Creating files with rand or pid to randomize the file
names is not adequate to protect the system.

We now use File::Temp to safely create the temporary files as needed.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>