[topic--man-fixes]: Fix man page comment schema
The commenting scheme used in the man pages was three consecutive single
quotes (`'''') at the beginning of a line. his works, but `gtroff' gives
a warning about an undefined macro (namely `'''), which is harmless, but
irritating.

To avoid this, this commit changes the schema to start the line with
`.\"' which causes the line to be treated as an undefined request and
thus ignored completely.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>